+++ Assignment2b/stack.cpp
@@ -3,18 +3,18 @@ second homework project
 */
 #include "stack.hpp"
 Stack::Stack(int size){
-    y = new int[size];
+    y = new double[size];
     top = -1;
 }
 Stack::~Stack(){
     delete[] y;
 }
-void Stack::add(int x){
+void Stack::add(double x){
     y[++top] = x;
 }
-int Stack::del(){
+double Stack::del(){
     return y[top--];
 }
-int Stack::take() const {
+double Stack::take() const {
     return y[top];
 }
